Moved redis class to __init__ for consistency.
